pyproject.toml
[project]
name = "gitgud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "discord-webhook>=1.4.1",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
